requirements.txt
tensorflow==2.3.1
pycocotools==2.0.2
numpy~=1.18.5
six~=1.15.0
matplotlib~=3.3.2
Pillow~=7.2.0
dataclasses~=0.6
sentencepiece~=0.1.91
scipy~=1.5.2
setuptools~=50.3.0
opencv-python~=4.4.0.42
PyYAML~=5.3.1
mock~=2.0.0
pandas~=1.1.3
scikit-image~=0.17.2
h5py~=2.10.0
contextlib2~=0.6.0.post1
ipython~=7.18.1
lxml~=4.5.2
networkx~=2.5
sklearn==0.23.2
